refactor: types for user and track
