+++ src/user/user-types.ts
@@ -1,0 +1,28 @@
+import { IsNotEmpty, IsString } from 'class-validator';
+
+export interface IUser {
+  id: string; // uuid v4
+  login: string;
+  password: string;
+  version: number; // integer number, increments on update
+  createdAt: number; // timestamp of creation
+  updatedAt: number; // timestamp of last update
+}
+
+export class ICreateUserDto {
+  @IsString()
+  @IsNotEmpty()
+  login: string;
+  @IsString()
+  @IsNotEmpty()
+  password: string;
+}
+
+export class IUpdatePasswordDto {
+  @IsString()
+  @IsNotEmpty()
+  oldPassword: string; // previous password
+  @IsString()
+  @IsNotEmpty()
+  newPassword: string; // new password
+}
