+++ src/user/user.service.ts
@@ -4,7 +4,7 @@ import {
   NotFoundException,
 } from '@nestjs/common';
 import { usersDb } from 'src/database/database';
-import { ICreateUserDto, IUpdatePasswordDto } from 'src/types/types';
+import { ICreateUserDto, IUpdatePasswordDto } from './user-types';
 import { v4 as uuidv4 } from 'uuid';
 import { validate } from 'uuid';
